+++ includes/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: skanna <skanna@student.42.fr>              +#+  +:+       +#+        */
+/*   By: deniseerjavec <deniseerjavec@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/08 17:03:55 by skanna            #+#    #+#             */
-/*   Updated: 2024/06/19 16:46:29 by skanna           ###   ########.fr       */
+/*   Updated: 2024/06/20 11:45:56 by deniseerjav      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -81,15 +81,19 @@ t_lst_env	*copy_list(t_lst_env *lst);
 
 //exec
 void		execution(t_mini *minishell);
+void	    cmd_exec(t_mini *mini, t_token *tmp);
 int			is_builtin(char *command);
 void		execute_builtin(t_mini *mini, int builtin, t_token *cur);
 int			init_fds(t_mini *mini);
 int			ft_dup(t_mini *mini, int i);
+int	        ft_dup_tubes(t_mini *mini, int i);
 int			fill_fd(t_mini *mini);
 
 //builtin fucntions
 void		echo_command(t_token *cur);
 void		export_command(t_mini *mini, t_token *cur);
+void        add_var_to_list(t_mini *mini, t_token *cur);
+char	    **list_to_tab(t_mini *mini);
 void		env_command(t_mini *minishell);
 void		exit_cmd(t_mini *minishell, char *input);
 void		pwd_cmd(t_mini *mini);
