+++ src/builtins/echo.c
@@ -6,32 +6,37 @@
 /*   By: deniseerjavec <deniseerjavec@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/15 15:52:38 by skanna            #+#    #+#             */
-/*   Updated: 2024/06/19 17:36:27 by deniseerjav      ###   ########.fr       */
+/*   Updated: 2024/06/20 10:43:47 by deniseerjav      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
+static void	check_after_dash(t_token *cur, int *n, int i)
+{
+	int	j;
+	
+	j = 1;
+	while (cur->cmd_tab[i][j])
+	{
+		if (cur->cmd_tab[i][j] != 'n')
+			break ;
+		j++;
+	}
+	if (cur->cmd_tab[i][j] == '\0')
+		++(*n);
+}
+
 static int	echo_opt(t_token *cur, int *n)
 {
 	int	i;
-	int	j;
+	
 
 	i = 1;
 	while (cur->cmd_tab[i] && ft_strncmp(cur->cmd_tab[i], "-n", 2) == 0)
 	{
 		if (ft_strlen(cur->cmd_tab[i]) > 2)
-		{
-			j = 1;
-			while (cur->cmd_tab[i][j])
-			{
-				if (cur->cmd_tab[i][j] != 'n')
-					break ;
-				j++;
-			}
-			if (cur->cmd_tab[i][j] == '\0')
-				++(*n);
-		}
+			check_after_dash(cur, n, i);
 		else
 			++(*n);
 		if (ft_strlen(cur->cmd_tab[i]) == 2 || *n > 0)
