+++ src/builtins/export/export.c
@@ -1,26 +1,38 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   pwd.c                                              :+:      :+:    :+:   */
+/*   export.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: deniseerjavec <deniseerjavec@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/05/15 15:52:20 by skanna            #+#    #+#             */
-/*   Updated: 2024/06/20 11:23:54 by deniseerjav      ###   ########.fr       */
+/*   Created: 2024/05/08 17:03:22 by skanna            #+#    #+#             */
+/*   Updated: 2024/06/20 11:12:22 by deniseerjav      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void	pwd_cmd(t_mini *mini)
+static void	print_export(t_lst_env *env)
 {
-	char	cwd[1024];
-	int		size;
-
-	size = 1024;
-	if (getcwd(cwd, size) == NULL)
+	while (env)
 	{
-		ft_error("pwd error", mini);
+		if (env->value)
+			printf("declare -x %s=\"%s\"\n", env->name, env->value);
+		else
+			printf("declare -x %s\n", env->name);
+		env = env->next;
 	}
-	printf("%s\n", cwd);
+}
+
+void	export_command(t_mini *mini, t_token *cur)
+{
+	char	*new_var;
+
+	new_var = cur->cmd_tab[1];
+	if (new_var)
+		add_var_to_list(mini, cur);
+	else
+		print_export(mini->export);
+	if (mini->error)
+		mini->exit_status = 1;
 }
