+++ src/exec/execute_builtin.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   execute_builtin.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: skanna <skanna@student.42.fr>              +#+  +:+       +#+        */
+/*   By: deniseerjavec <deniseerjavec@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/08 17:03:22 by skanna            #+#    #+#             */
-/*   Updated: 2024/06/19 17:00:05 by skanna           ###   ########.fr       */
+/*   Updated: 2024/06/20 11:25:01 by deniseerjav      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,15 +46,3 @@ void	execute_builtin(t_mini *mini, int builtin, t_token *cur)
 	else if (builtin == 5)
 		cd_cmd(mini);
 }
-
-
-// void	execute_commands(t_mini *mini, char *command, t_token *cur)
-// {
-// 	int	builtin;
-
-// 	builtin = is_builtin(command);
-// 	if (builtin > 0)
-// 		execute_builtin(mini, builtin, cur);
-// 	else
-// 		cmd_exec(mini, cur);
-// }
