+++ src/exec/execution.c
@@ -3,40 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   execution.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: skanna <skanna@student.42.fr>              +#+  +:+       +#+        */
+/*   By: deniseerjavec <deniseerjavec@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/08 17:03:22 by skanna            #+#    #+#             */
-/*   Updated: 2024/06/19 18:02:16 by skanna           ###   ########.fr       */
+/*   Updated: 2024/06/20 11:58:11 by deniseerjav      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-static char	**pars_path(t_mini *mini)
-{
-	char	*path;
-	char	**split_paths;
-
-	path = get_env_value(mini->env, "PATH");
-	if (path == NULL)
-		return (NULL);
-	split_paths = ft_split(path, ':');
-	free(path);
-	return (split_paths);
-}
-
-static char	*join_path(char *cmd, char *path)
-{
-	char	*joint_a;
-	char	*joint_b;
-
-	joint_a = ft_strjoin_char(path, '/');
-	if (joint_a == NULL)
-		return (NULL);
-	joint_b = ft_strjoin(joint_a, cmd);
-	free(joint_a);
-	return (joint_b);
-}
 
 static void	close_fd_and_wait(t_mini *mini)
 {
@@ -60,58 +34,31 @@ static void	close_fd_and_wait(t_mini *mini)
 	mini->exit_status = last_exit_status;
 }
 
-static void	cmd_exec(t_mini *mini, t_token *tmp)
+static void	child_pid(t_mini *mini, t_token	*tmp, int i)
 {
-	int		i;
-	char	**paths;
-	char	*path_with_token;
-
-	i = 0;
-	paths = pars_path(mini);
-	if (paths == NULL)
-		return (ft_error("Memory allocation error in execution", mini));
-	while (paths[i])
-	{
-		path_with_token = join_path(tmp->cmd_tab[0], paths[i]);
-		if (!path_with_token)
-			return (free(paths));
-		if (access(path_with_token, X_OK) == 0)
-		{
-			if (execve(path_with_token, tmp->cmd_tab, mini->env_char) == -1)
-			{
-				ft_error("execve Error", mini);
-				exit(1);
-			}
-		}
-		free(path_with_token);
-		i++;
-	}
-	printf("%s :command not found\n", tmp->cmd_tab[0]);
-	free(paths);
-	//ft_error("command not found", mini);
-	exit(127);
-}
-
-void	execution(t_mini *mini)
-{
-	t_token	*tmp;
-	int		i;
 	int		builtin;
 
-	if (init_fds(mini) != 0)
-		return ;
-	if (fill_fd(mini) != 0)
-		return ;
-	if (mini->cmd_count <= 0)
-		return ;
-	tmp = mini->token;
-	while (tmp->type != COMMAND)
-		tmp = tmp->next;
+	if (ft_dup(mini, i) != 0)
+		exit (1);
 	builtin = is_builtin(tmp->cmd_tab[0]);
-	if (mini->cmd_count == 1 && builtin > 0)
-		return (execute_builtin(mini, builtin, mini->token));
+	if (builtin > 0)
+		execute_builtin(mini, builtin, tmp);
+	else
+		cmd_exec(mini, tmp);
+	if (mini->error)
+	{
+		clean_minishell(mini);
+		exit (1);
+	}
+	clean_minishell(mini);
+	exit (0);
+}
+
+static void	exec_in_child(t_mini *mini, t_token	*tmp)
+{
+	int		i;
+	
 	i = 0;
-	tmp = mini->token;
 	while (tmp)
 	{
 		if (tmp->type == COMMAND)
@@ -122,22 +69,7 @@ void	execution(t_mini *mini)
 			if (mini->pid[i] < 0)
 				return (ft_error("Fork error", mini));
 			if (mini->pid[i] == 0)
-			{
-				if (ft_dup(mini, i) != 0)
-					exit (1);
-				builtin = is_builtin(tmp->cmd_tab[0]);
-				if (builtin > 0)
-					execute_builtin(mini, builtin, tmp);
-				else
-					cmd_exec(mini, tmp);
-				if (mini->error)
-				{
-					clean_minishell(mini);
-					exit (1);
-				}
-				clean_minishell(mini);
-				exit (0);
-			}
+				child_pid(mini, tmp, i);
 			if (i > 0)
 			{
 				close(mini->tube[i - 1][0]);
@@ -147,5 +79,38 @@ void	execution(t_mini *mini)
 		}
 		tmp = tmp->next;
 	}
+}
+
+
+static int builtin_in_parent(t_mini *mini)
+{
+	t_token	*tmp;
+	int		builtin;
+	
+	tmp = mini->token;
+	while (tmp->type != COMMAND)
+		tmp = tmp->next;
+	builtin = is_builtin(tmp->cmd_tab[0]);
+	if (mini->cmd_count == 1 && builtin > 0)
+		return (execute_builtin(mini, builtin, mini->token), 1);
+	return (0);
+}
+
+void	execution(t_mini *mini)
+{
+	t_token	*tmp;
+
+	if (init_fds(mini) != 0)
+		return ;
+	if (fill_fd(mini) != 0)
+		return ;
+	if (mini->cmd_count <= 0)
+		return ;
+	if (builtin_in_parent(mini) != 0)
+		return ;
+	tmp = mini->token;
+	exec_in_child(mini, tmp);
+	if (mini->error)
+		return ;
 	close_fd_and_wait(mini);
 }
